update: link scripts

+++ hammerspoon/link.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
 #################################################################
-#   Copyright (C) 2016 Sean Guo. All rights reserved.
+#   Copyright (C) 2017 Sean Guo. All rights reserved.
 #
-#	> File Name:        < ln-link.sh >
+#	> File Name:        < link.sh >
 #	> Author:           < Sean Guo >
 #	> Mail:             < iseanxp+code@gmail.com >
 #	> Created Time:     < 2016/03/26 >
-#	> Last Changed:     < 2016/11/13 >
+#	> Last Changed:     < 2017/01/21 >
 #	> Description:      hammerspoon config automatic deployment
 #################################################################
 
@@ -17,18 +17,19 @@ target="$base_dir/$dotfile"
 modules_dir="$base_dir/modules"
 modules_files="`ls -1 modules`"
 
-rm -r $modules_dir
+if [ -d $modules_dir ];then
+  rm -r $modules_dir
+fi
 mkdir -p $base_dir $modules_dir
 
-# backuo old config file
+# backup old config file
 if [ -f "$target" ];then
-  mv "$target" "$target".old
+  mv "$target" "$target.old"
 fi
 
 # make soft link
 ln -s `pwd`/$dotfile "$target"
-
 for lua in $modules_files
 do
-  ln -sf "`pwd`/modules/$lua" $modules_dir
+  ln -s "`pwd`/modules/$lua" $modules_dir
 done
